third_party/blink/web_tests/external/wpt/wasm/jsapi/function/constructor.tentative.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] length
  assert_equals: WebAssembly.Function length should be 2 expected 2 but got 1
Harness: the test ran to completion.
